requirements.txt
PyQt6>=6.6.0
mss>=9.0.0
Pillow>=10.0.0
keyboard>=0.13.5
pywin32>=306
